chore: Update build.yml workflow to include Docker image push with all tags
This commit updates the build.yml workflow to include a new step for pushing the Docker image with all tags to the Docker registry. Instead of pushing individual tags, the `docker push --all-tags` command is used to push all tags associated with the repository. This change ensures that all versions of the Docker image are pushed to the registry for deployment.