chore: Update Amazon ECS task definition to use latest version of todo-csharp-sql image
This commit updates the Amazon ECS task definition to use the latest version (v0.0.35) of the todo-csharp-sql image. By updating the image, we ensure that the ECS task is running the most up-to-date version of the application, incorporating any bug fixes or feature enhancements. This change improves the reliability and functionality of the ECS task.