chore: Update deploy.yml workflow to use Google Cloud Run instead of GKE
This commit updates the deploy.yml workflow to use the Google Cloud Run action instead of the Google Kubernetes Engine (GKE) action. The GCP_GKE_CLUSTER, GCP_GKE_REGION, and GCP_PROJECT_ID variables are no longer needed and have been replaced with the GCP_RUN_SERVICE and inputs.version variables. This change aligns the deployment process with the preferred platform and improves the clarity and maintainability of the workflow configuration.