chore: Update Amazon ECS task definition to use new Docker image
This commit updates the Amazon ECS task definition in the `amazon-ecs-task-definition.json` file to use a new Docker image `docker.io/thiagolunardi/todo-csharp-sql:latest` for the `todo-csharp-sql` container. The previous image `public.ecr.aws/docker/library/httpd:latest` is replaced. This change improves the application's functionality and ensures that the correct image is used for the container.